Got shared caching working as a library, much faster now
