+++ src/compat/compat.wake
@@ -18,4 +18,4 @@ package build_wake
 from wake import _
 
 target compat variant =
-    src here variant Nil Nil
+    src here variant Nil Nil Nil
