+++ src/job-cache/hash.h
