+++ src/job-cache/logging.h
@@ -21,8 +21,6 @@
 #define _XOPEN_SOURCE 700
 #define _POSIX_C_SOURCE 200809L
 
-#pragma once
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
