+++ src/json/json.wake
@@ -18,4 +18,4 @@ package build_wake
 from wake import _
 
 target json variant =
-    src here variant (util, Nil) Nil
+    src here variant (util, Nil) Nil Nil
