+++ src/optimizer/optimizer.wake
@@ -18,4 +18,4 @@ package build_wake
 from wake import _
 
 target optimizer variant =
-    src here variant (dst, types, util, Nil) (gmp, Nil)
+    src here variant (dst, types, util, Nil) (gmp, Nil) Nil
