+++ src/parser/parser.wake
@@ -18,4 +18,4 @@ package build_wake
 from wake import _
 
 target parser variant =
-    src here variant (json, util, Nil) (re2, utf8proc, Nil)
+    src here variant (json, util, Nil) (re2, utf8proc, Nil) Nil
