+++ src/wcl/optional.h
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+#pragma once
+
 #include <cstdint>
 #include <utility>
